api/pyproject.toml
[project]
name = "db-host-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "backboard-sdk>=0.1",
    "distributed-sqlite>=0.3.0",
    "boto3>=1.35",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
    "httpx>=0.28",
    "pydantic[email]>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
